--- front/components/APropos.vue
@@ -1,21 +1,0 @@
-<template>
-    <div class="flex justify-center">
-        <div class="flex flex-col w-1/3 pt-64 items-center text-center">
-            <h1 class="text-6xl">A propos</h1>
-            <p class="text-xl my-6">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Risus justo ac porttitor fringilla non venenatis purus. Vel auctor at tincidunt quisque commodo vehicula velit. Mattis mauris egestas nisl amet ridiculus. Quis neque, aliquam enim imperdiet sed risus arcu mus. Suspendisse turpis eget mattis etiam lectus ut. Quis quis gravida viverra diam magna vitae pretium massa id. Dignissim integer ac ultrices ipsum morbi. Habitant dolor, nullam in tellus nunc, auctor non. In malesuada adipiscing pulvinar placerat viverra molestie arcu.</p>
-            <h1 class="text-6xl">Mentions légales</h1>
-            <p class="text-xl my-6">Yo</p>
-            <h1 class="text-6xl">Crédits</h1>
-            <p class="text-xl my-6">Coucou c'est nous</p>
-        </div>
-    </div>
-</template>
-
-<script>
-export default {
-    name: "APropos",
-}
-</script>
-
-<style>
-</style>
